requirement.txt
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.0
dadjokes==1.3.2
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.2
fonttools==4.41.1
fqdn==1.5.1
fsspec==2023.6.0
futures==3.0.5
goslate==1.5.4
huggingface-hub==0.16.4
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jaraco.context==4.3.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==9.1.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
newsapi-python==0.2.7
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.1
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.17.0
pluggy==1.2.0
polars==0.19.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
PyDictionary==2.0.1
pydotplus==2.0.2
Pygments==2.16.1
pyparsing==3.0.9
PyPDF2==3.0.1
pypiwin32==223
pytest==7.4.0
pytest-benchmark==4.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.4.0
randfacts==0.20.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
safetensors==0.3.2
scikit-learn==1.4.1.post1
scipy==1.11.2
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.13.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
widgetsnbextension==4.0.8
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
xmltodict==0.13.0
